Move container run files out of library

+++ SWARMRDSClientCore/run_swarm_container.sh
@@ -13,4 +13,11 @@ else
     mkdir .cache
 fi
 
+# Check that the User has input the name of the Docker image to run
+if [ -z "$1" ]
+then
+    echo "Please input the name of the Docker image to run."
+    exit 1
+fi
+
 docker run -it --rm --gpus=all --runtime=nvidia --network=host -e "ROS_IP_ADDRESS=$2" -v $pwd/.cache:/home/airsim_user/SWARMCore/core/.cache $1
